blfoat16 to float16

+++ finetuning/conditional_generation/olmo_evaluate.py
@@ -62,7 +62,7 @@ def load_model_and_tokenizer(model_args, args):
         load_in_4bit=True,
         bnb_4bit_quant_type="nf4",
         bnb_4bit_use_double_quant=True,
-        bnb_4bit_compute_dtype=torch.bfloat16,
+        bnb_4bit_compute_dtype=torch.float16,
     )
 
     base_model = AutoModelForCausalLM.from_pretrained(
